Make the struct bigger, to ensure it is returned by struct return.
llvm-svn: 50037

+++ llvm/test/CFrontend/2007-05-07-NestedStructReturn.c
@@ -1,6 +1,6 @@
 // RUN: %llvmgcc %s -S -fnested-functions -o - | grep {sret *%agg.result}
 
-struct X { int m, n, o, p; };
+struct X { long m, n, o, p; };
 
 struct X p(int n) {
   struct X c(int m) {
